+++ test/helpers.js
@@ -4,24 +4,31 @@ var isNode = typeof window === 'undefined'
 
 function stubFetchAPI(spy, data, dataType) {
   data = data || '{}';
+
+  var response = { status: 200 };
+  if (dataType) {
+    response[dataType] = spy;
+  }
+  else {
+    response.json = spy;
+    response.text = spy;
+  }
+
   return sinon.stub(global, 'fetch', function(path, options) {
     return {
-      then: function(callback1) {
-        var response = {};
-        if (dataType) {
-          response[dataType] = spy;
-        }
-        else {
-          response.json = spy;
-          response.text = spy;
-        }
-        callback1.call(this, response);
+      then: function(callback0) {
+        callback0.call(this, response);
         return {
-          then: function(callback2) {
-            callback2.call(this, data);
+          then: function(callback1) {
+            callback1.call(this, response);
             return {
-              catch: function(callback3) {
-                callback3.call(this, new Error(''));
+              then: function(callback2) {
+                callback2.call(this, data);
+                return {
+                  catch: function(callback3) {
+                    callback3.call(this, new Error(''));
+                  }
+                }
               }
             }
           }